+++ .prettierrc
@@ -7,7 +7,15 @@
     "@ianvs/prettier-plugin-sort-imports",
     "prettier-plugin-tailwindcss"
   ],
-  "importOrder": ["react", "", "@/components", ""],
+  "importOrder": [
+    "react",
+    "",
+    "@/components",
+    "",
+    "@/interfaces",
+    "",
+    "@/hooks"
+  ],
   "importOrderParserPlugins": ["typescript", "jsx", "decorators-legacy"],
   "importOrderTypeScriptVersion": "5.0.0",
   "importOrderCaseSensitive": false
