+++ src/__test__/mockData/handlers.ts
@@ -4,8 +4,9 @@ import {
   MockDetailedCountry,
 } from './countries.mock';
 import { BASE_API_URL } from '@/constants';
-import type { ICountry, IDetailedCountry } from '@/interfaces';
 import { http, HttpResponse } from 'msw';
+
+import type { ICountry, IDetailedCountry } from '@/interfaces';
 
 const NETWORK_DELAY = 10;
