+++ src/components/theme-changer/themeContext.ts
@@ -1,5 +1,6 @@
-import type { ThemeContextType } from '@/interfaces';
 import { createContext } from 'react';
+
+import type { ThemeContextType } from '@/interfaces';
 
 export const ThemeContext = createContext<ThemeContextType | undefined>(
   undefined
