+++ src/pages/home/Home.tsx
@@ -1,4 +1,3 @@
-import type { ICountry } from '@/interfaces';
 import clsx from 'clsx';
 import { useState } from 'react';
 import { Outlet, useLoaderData, useParams } from 'react-router';
@@ -10,6 +9,8 @@ import {
   Pagination,
   SearchForm,
 } from '@/components';
+
+import type { ICountry } from '@/interfaces';
 
 interface LoaderData {
   countries: ICountry[] | [];
